+++ super/S_Login/bot_new.py
@@ -1,6 +1,6 @@
 """
 
-from newapi.super.login_bots.bot_new import LOGIN_HELPS
+from newapi.super.S_Login.bot_new import LOGIN_HELPS
 
 Exception:{'login': {'result': 'Failed', 'reason': 'You have made too many recent login attempts. Please wait 5 minutes before trying again.'}}
 
@@ -14,14 +14,14 @@ import requests
 from http.cookiejar import MozillaCookieJar
 
 from newapi import printe
-from newapi.super.login_bots.cookies_bot import get_file_name, del_cookies_file
+from newapi.super.S_Login.cookies_bot import get_file_name, del_cookies_file
 from newapi.except_err import exception_err
-from newapi.super.login_bots.params_help import PARAMS_HELPS
+from newapi.super.S_Login.params_help import PARAMS_HELPS
 
 # import mwclient
 
 # from mwclient.client import Site
-from newapi.super.login_bots.mwclient.client import Site
+from newapi.super.S_Login.mwclient.client import Site
 
 # cookies = get_cookies(lang, family, username)
 users_by_lang = {}
