+++ super/super_login.py
@@ -24,9 +24,9 @@ from newapi.super.bots.handel_errors import HANDEL_ERRORS
 from newapi.except_err import warn_err
 
 if "nomwclient" in sys.argv:
-    from newapi.super.login_bots.bot import LOGIN_HELPS
+    from newapi.super.S_Login.bot import LOGIN_HELPS
 else:
-    from newapi.super.login_bots.bot_new import LOGIN_HELPS
+    from newapi.super.S_Login.bot_new import LOGIN_HELPS
 
 file_name = os.path.basename(__file__)
 print_test = {1: False}
